% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{approximateEnu}
\alias{approximateEnu}
\title{approximateEnu}
\usage{
approximateEnu(n, K, degCutoff, Y, clique, clique_size)
}
\arguments{
\item{n}{number of vertices in graph}

\item{K}{number of cliques}

\item{degCutoff}{a number}

\item{Y}{adjacent matrix}

\item{clique}{a matrix}

\item{clique_size}{a number}
}
\value{
a number
}
\description{
approximateEnu
}
