% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Romano.R
\name{Bootstrap_Romano}
\alias{Bootstrap_Romano}
\title{Bootstrap function}
\usage{
Bootstrap_Romano(
  Y,
  cliques,
  clique_size,
  K,
  E_nu,
  J,
  lambdaHat,
  alpha,
  m,
  geometry,
  kappa,
  rate,
  zeta
)
}
\arguments{
\item{Y}{a matrix}

\item{cliques}{cliques of matrix}

\item{clique_size}{size od clique}

\item{K}{number of cliques}

\item{E_nu}{a nu ber}

\item{J}{a matrix}

\item{lambdaHat}{estimate}

\item{alpha}{a nmber}

\item{m}{size - 1}

\item{geometry}{type of space}

\item{kappa}{curvature of latent space}

\item{rate}{rate used when sub-sampling}

\item{zeta}{scaling in front of distances}
}
\value{
list of True/False, p-value
}
\description{
Bootstrap function
}
