% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Curvature.R
\name{Estimate_Curvature_Hyperbolic}
\alias{Estimate_Curvature_Hyperbolic}
\title{Estimate_Curvature_Hyperbolic Function}
\usage{
Estimate_Curvature_Hyperbolic(dHat, K, a, b, index)
}
\arguments{
\item{dHat}{a matrix}

\item{K}{number of cliques}

\item{a}{lower bound}

\item{b}{upper bound}

\item{index}{a index number}
}
\value{
minimum kappaVec
}
\description{
Estimate_Curvature_Hyperbolic Function
}
