% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{EstimatePhat}
\alias{EstimatePhat}
\title{EstimatePhat}
\usage{
EstimatePhat(k, kPrime, K, n, c, Y)
}
\arguments{
\item{k}{a number}

\item{kPrime}{a number}

\item{K}{number of cliques}

\item{n}{clique size times number of cliques}

\item{c}{clique}

\item{Y}{adjacency matrix for subgraph induced by cliques}
}
\value{
list of two number
}
\description{
EstimatePhat
}
