% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{findCliques_GivenY}
\alias{findCliques_GivenY}
\title{findCliques_GivenY}
\usage{
findCliques_GivenY(Y, K, n, clique_size, E_nu, numTaken, t)
}
\arguments{
\item{Y}{adjacency matrix}

\item{K}{number of cliques}

\item{n}{size of graph}

\item{clique_size}{clique size}

\item{E_nu}{a number}

\item{numTaken}{determines how many nodes to take in one group to find cliques}

\item{t}{number of edges used in the "almost clique" to compute FE expectation.}
}
\value{
list of matrix dHat and matrix finalCliques
}
\description{
findCliques_GivenY
}
