% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Hyperbolic.R
\name{Generate_Hyperbolic}
\alias{Generate_Hyperbolic}
\title{Generate_Hyperbolic function}
\usage{
Generate_Hyperbolic(n, dim, scale, kappa)
}
\arguments{
\item{n}{number of cliques}

\item{dim}{dimension of latent space}

\item{scale}{scale used when generating points in hyperbolic space.}

\item{kappa}{curvature of latent space.}
}
\value{
muVec and Droups
}
\description{
Generate_Hyperbolic function
}
