% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{GenerateGraph}
\alias{GenerateGraph}
\title{GenerateGraph}
\usage{
GenerateGraph(n, probs)
}
\arguments{
\item{n}{siz of graph}

\item{probs}{probability to create an edge}
}
\value{
matrix Y
}
\description{
GenerateGraph
}
