% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{GenerateY}
\alias{GenerateY}
\title{GenerateY}
\usage{
GenerateY(K, n, probs, clique_size, E_nu, numTaken, t)
}
\arguments{
\item{K}{number of cliques}

\item{n}{size of graph}

\item{probs}{probability to create an edge}

\item{clique_size}{size of clique}

\item{E_nu}{a number}

\item{numTaken}{determines how many nodes to take in one group to find cliques}

\item{t}{number of edges used in the "almost clique" to compute FE expectation}
}
\value{
Y, pHat, dHat, finalCliques, E_nu
}
\description{
GenerateY
}
