% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Hyperbolic.R
\name{Make_D_FromMu_Hyperbolic}
\alias{Make_D_FromMu_Hyperbolic}
\title{Make_D_FromMu_Hyperbolic function}
\usage{
Make_D_FromMu_Hyperbolic(n, p, K, muVec, kappa, delta)
}
\arguments{
\item{n}{size of graph.}

\item{p}{dimension of latent space}

\item{K}{number of cliques}

\item{muVec}{a matrix}

\item{kappa}{curvature of latent space}

\item{delta}{ontrols the spread of the nodes around their group centers in spherical LS}
}
\value{
matrix D
}
\description{
Make_D_FromMu_Hyperbolic function
}
