% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Euclidean_D.R
\name{Make_Euclidean_D}
\alias{Make_Euclidean_D}
\title{Make_Euclidean_D function}
\usage{
Make_Euclidean_D(K, n, sigma, p)
}
\arguments{
\item{K}{number of cliques}

\item{n}{size of graph}

\item{sigma}{controls the spread of the points in Euclidean space}

\item{p}{dimension of latent space}
}
\value{
list of matrix DGroups and matrix D
}
\description{
Make_Euclidean_D function
}
