% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_D_Spherical.R
\name{Make_Separated_D_Spherical_3D}
\alias{Make_Separated_D_Spherical_3D}
\title{Make_Separated_D_Spherical_3D function}
\usage{
Make_Separated_D_Spherical_3D(K, n, kappa, ub_theta, ub_phi, delta)
}
\arguments{
\item{K}{number of cliques}

\item{n}{size of graph}

\item{kappa}{curvature of latent space}

\item{ub_theta}{upper bound for distribution of theta angle for points in  spherical LS}

\item{ub_phi}{upper bound for distribution of phi angle for points in  spherical LS}

\item{delta}{controls the spread of the nodes around their group centers in spherical LS}
}
\value{
list of matrix DGroups and matrix D
}
\description{
Make_Separated_D_Spherical_3D function
}
