% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{PickCliquesFeasible}
\alias{PickCliquesFeasible}
\title{PickCliquesFeasible}
\usage{
PickCliquesFeasible(n, K, Y, clique, numSamples, clique_size, t)
}
\arguments{
\item{n}{number of vertices of the graph}

\item{K}{a number}

\item{Y}{adjacent matrix}

\item{clique}{list of lists}

\item{numSamples}{number of cliques}

\item{clique_size}{clique size}

\item{t}{number of edges used in the "almost clique" to compute FE expectation}
}
\value{
list of optimal indices, minimum objFun, True/False of whether min(objFun) < 100, dHat, cliques, Enu
}
\description{
PickCliquesFeasible
}
