% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{Return_Classification}
\alias{Return_Classification}
\title{Return_Classification}
\usage{
Return_Classification(
  Y,
  clique_size,
  K,
  m,
  mode,
  computeRate,
  finalCliques,
  dHat,
  E_nu = 1
)
}
\arguments{
\item{Y}{adjacent matrix}

\item{clique_size}{a number}

\item{K}{number of cliques}

\item{m}{clique_size - 1}

\item{mode}{current not use, default set to "Not Data"}

\item{computeRate}{current not use, default set to "False"}

\item{finalCliques}{list of lists}

\item{dHat}{matrix}

\item{E_nu}{defalut set to 1}
}
\value{
list of classification_max, p_E, p_S, p_H, kappaHat_S, kappaHat_H, a, b
}
\description{
Return_Classification
}
