% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplementary.R
\name{sort_cliques}
\alias{sort_cliques}
\title{sort_cliques function}
\usage{
sort_cliques(cliques_adjusted)
}
\arguments{
\item{cliques_adjusted}{a list of lists}
}
\value{
reordered cliques
}
\description{
sort_cliques function
}
