% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type1_Power.R
\name{squeeze}
\alias{squeeze}
\title{squeeze function}
\usage{
squeeze(arr)
}
\arguments{
\item{arr}{array with one or more dimensions of size 1}
}
\value{
array with all single-dimensional (length 1) entries removed from its shape
}
\description{
squeeze function
}
